Structure container moved to utils.  get_FIGURE moved to plotting.py.
