+++ reflexible/plotting.py
@@ -9,7 +9,8 @@ import matplotlib.pyplot as plt
 from mpl_toolkits import basemap
 
 # local imports
-from reflexible import mapping as mp
+import reflexible.mapping as mp
+from reflexible import Structure
 
 
 def plot_at_level(H, data, level=1,
@@ -121,6 +122,80 @@ def plot_totalcolumn(H, data=None,
     return FIGURE
 
 
+def get_FIGURE(fig=None, ax=None, m=None, map_region=None,
+               getm=True, map_par=None, fig_par=None,
+               image=None):
+    """Returns a matplotlib figure based on the parameters.
+
+    The idea is that I create a :class:`Structure` that contains the figure,
+    ax, and m instance. I also add a field for "indices".
+    This is so as to be able to reuse figures. This saves a huge amount of
+    time, as creating then basemap instance can be time consuming.
+
+    .. note::
+        Generally you won't use this function directly.
+
+    USAGE::
+
+        > FIG = get_FIGURE()
+        or
+        > FIG = get_FIGURE(map_region='polarcat')
+
+    Returns
+       This will return the "FIG" object, which has attributes: `fig`, `ax`,
+       `m`, and `indices`. The indices are used for deleting lines, texts,
+       collections, etc. if and when we are reusing the figure instance. The
+       indices basically give us a reference to the *empty* map, so we can
+       delete lines without losing meridians or parallels for example.
+
+         ============      ======================================
+         keys              description
+         ============      ======================================
+         fig               A pyplot.fig instance, use
+                           plt.figure(FIG.fig.number) to make the
+                           fig active (for example to use
+                           plt.savefig('filename.png')
+         m                 The basemap instance so you can do:
+                           x,y = FIG.m(lon,lat)
+         ax                The axes
+         indices           with index for texts, images, lines,
+                           and collections
+         ============      ======================================
+
+    """
+    FIGURE = Structure()
+
+    if getm and m is None:
+        if image:
+            fig, m = mp.get_base_image(image, map_region=map_region,
+                                       map_par=map_par, fig_par=fig_par)
+        else:
+            fig, m = mp.get_base1(map_region=map_region, map_par=map_par,
+                                  fig_par=fig_par, fig=fig)
+            FIGURE.fig = fig
+            FIGURE.m = m
+            FIGURE.ax = fig.gca()
+    else:
+        FIGURE.m = m
+
+    if fig is None:
+        FIGURE.fig = plt.figure()
+        fig = FIGURE.fig
+    else:
+        FIGURE.fig = fig
+
+    FIGURE.ax = ax if ax is not None else fig.gca()
+
+    FIGURE.indices = Structure()
+    FIGURE.indices.texts = len(FIGURE.ax.texts)
+    FIGURE.indices.images = len(FIGURE.ax.images)
+    FIGURE.indices.collections = len(FIGURE.ax.collections)
+    FIGURE.indices.lines = len(FIGURE.ax.lines)
+
+    print("Using figure: %s" % FIGURE.fig.number)
+    return FIGURE
+
+
 def plot_sensitivity(H, data,
                      data_range=None,
                      units='ns m^2 / kg',
@@ -136,23 +211,24 @@ def plot_sensitivity(H, data,
                      map_par=None,
                      fig_par=None,
                      cax_title=None,
-                     method='contourf', lsmask=False):
+                     method='contourf',
+                     lsmask=False):
     """ plot_sensitivity: core function for plotting FLEXPART output.
 
     Usage::
-        > FIG = plot_sensitivity(H,data,*kwargs)
+        FIG = plot_sensitivity(H,data,*kwargs)
 
     This returns the FIGURE object, and plots the sensitivity from the data
     contained in the "D" array.
 
     Inputs
-       H : a :class:`Header` instance for a FLEXPART run.
-       data : a 2d data array containing the sensitivity values to plot,
-         this can be extracted from a
-         grid instance (see :func:`readgridV8` and :func:`get_slabs`)
+        H : a :class:`Header` instance for a FLEXPART run.
+        data : a 2d data array containing the sensitivity values to plot,
+            this can be extracted from a grid instance
+            (see :func:`readgridV8` and :func:`get_slabs`)
 
     Returns
-      A "mapping.py" ``FIGURE`` object.
+        A "mapping.py" ``FIGURE`` object.
 
     Arguments
 
@@ -169,7 +245,7 @@ def plot_sensitivity(H, data,
       datainfo_str          A string for labeling the scale bar.
       plottitle             Title for the plot.
       rel_i                 Release index to plot from the data array
-      map_region                A map_region specified in mapping.py
+      map_region            A map_region specified in mapping.py
       dropm                 Force creation of a new basemap instance
       overlay               Force removal of previous figure elements.
       transform             For use with imshow method, if your data is not
@@ -177,7 +253,7 @@ def plot_sensitivity(H, data,
                             the data to the basemap projection.
       log                   Create a logarithmic color scale.
       FIGURE                A FIGURE instance from mapping module get_FIGURE
-      map_par                A Structure of paramters to be passed to the
+      map_par               A Structure of parameters to be passed to the
                             basemap class when creating an instance.
       method                The method to use for plotting array data. May be
                             one of: [pcolormesh], imshow, or contourf
@@ -186,7 +262,7 @@ def plot_sensitivity(H, data,
 
     .. todo::
         A lot!! There are some problems here and it is sensitive to options.
-        lsmask = True seems to only work with certain projections (POLARCAT)
+        lsmask = True seems to only work with certain projections (polarcat).
 
     .. note::
         This is the primary main function for creating plots of flexpart
@@ -203,13 +279,12 @@ def plot_sensitivity(H, data,
     assert method in methods, "method keyword must be one of: %s" % methods
 
     if FIGURE is None:
-        FIGURE = mp.get_FIGURE(map_region=map_region,
-                               map_par=map_par, fig_par=fig_par)
-    else:
-        if FIGURE.m is None:
-            FIGURE = mp.get_FIGURE(fig=FIGURE.fig, ax=FIGURE.ax,
-                                   map_region=map_region,
-                                   map_par=map_par, fig_par=fig_par)
+        FIGURE = get_FIGURE(map_region=map_region,
+                            map_par=map_par, fig_par=fig_par)
+    elif FIGURE.m is None:
+        FIGURE = get_FIGURE(fig=FIGURE.fig, ax=FIGURE.ax,
+                            map_region=map_region,
+                            map_par=map_par, fig_par=fig_par)
 
     if overlay is False:
         del FIGURE.ax.images[FIGURE.indices.images:]
@@ -290,7 +365,6 @@ def plot_sensitivity(H, data,
 
     if log:
         clevs = _gen_log_clevs(dat_min, dat_max)
-
     else:
         clevs = [i for i in
                  np.arange(dat_min, dat_max, (dat_max - dat_min) / 100)]
@@ -318,6 +392,7 @@ def plot_sensitivity(H, data,
     colmap.set_over(color='k', alpha=0.8)
     # Plotting METHODS (pcolormesh now default, imshow is smoother)
     # print(topodat.max(), topodat.min(), topodat.shape)
+    # ERROR: tododat might be used uninitialized
     if method == 'imshow':
         im = m.imshow(topodat, cmap=colmap, zorder=-1,
                       norm=mpl.colors.LogNorm(vmin=clevs[0],
@@ -471,8 +546,8 @@ def plot_trajectory(H, T, rel_i, FIGURE=None,
       days_back             For how many days back should the labels be
                             shown? [20]
       cbar2                 [True] draws the scale bar as a second axis.
-      cbar2_title            Optional argument to overide the cbar title.
-      map_par                A Structure of mapping parameters to pass
+      cbar2_title           Optional argument to overide the cbar title.
+      map_par               A Structure of mapping parameters to pass
                             to the basemap instance if desired.
       =============         =============================================
 
@@ -490,8 +565,8 @@ def plot_trajectory(H, T, rel_i, FIGURE=None,
 
     # Set up the FIGURE
     if FIGURE is None:
-        FIGURE = mp.get_FIGURE(map_region=map_region, map_par=map_par)
-    # #Get fig info and make active
+        FIGURE = get_FIGURE(map_region=map_region, map_par=map_par)
+    # Get fig info and make active
     fig = FIGURE.fig
     m = FIGURE.m
     ax = FIGURE.fig.axes[0]
@@ -502,7 +577,7 @@ def plot_trajectory(H, T, rel_i, FIGURE=None,
     trjs = T['Trajectories']
     rel = rel_i + 1  # account for zero indexing
 
-    # #extract only releases of interest
+    # extract only releases of interest
     t = trjs[np.where(trjs[:, 0] == rel), :][0]
 
     # Get the data for the days_back we're interested in
